2010-01-20	Joel Sherrill <joel.sherrill@oarcorp.com>
	* sp43/init.c: Fix warning.

+++ testsuites/sptests/sp43/init.c
@@ -86,8 +86,8 @@ void change_name(
   else {
     printf( "(" );
     for (c=newName ; *c ; ) {
-       if (isprint(*c)) printf( "%c", *c );
-       else             printf( "0x%02x", *c );
+       if (isprint((int)*c)) printf( "%c", *c );
+       else                  printf( "0x%02x", *c );
        c++;
        if ( *c )
          printf( "-" );
